requirement.txt
pandas==2.0.3
numpy==1.25.1
scikit-learn==1.3.2
Pillow==10.1.0
